requirements.txt
streamlit
pandas
plotly
folium
streamlit_folium
dnspython
pymongo
bcrypt
jwt
xlsxwriter
pytz